Better way of dealing with explicit sharing
Introduce the explicit set of keys in the datatype, with an invariant that all
known keys must be listed. This makes everything more uniform, and moreover I
was worried that we might have key environments with unused keys (which might
be useful for other reasons, perhaps they are old keys).