Prototype: fix package uploading
In the first attempt we listed the hash of the package directly in the
snapshot, but this is incorrect. We need an additional level of indirection:
the snapshot lists the targets.json which is specific for this package version,
which in turn list the hash of the package tarball itself.