Fix big in check for updates
We have to make sure that any files we download are only cached when the entire
check for updates completes succesfully.

In addition, we have to make sure to consider the root info as unchanged if we
didn't have a previous snapshot, or else the whole thing would loop
indefinitely.