Be more explicit about files with multiple formats
This makes it possible to be more precise about filepaths and filesizes for
different formats, and makes it easier to start making decisions about which
formats to download. (Might prefer, say, compressed on initial download, but
then uncompressed on subsequent downloads when using incremental updates.)

However, it's a bit dynamically typed right now, will see if that's fixable.