Remove Ord constraint from Some
This was previously necessary because we had a set of (Some Key), but now our
key environment goes from KeyIds instead, and we no longer need the Ord
constraint. This makes life a _lot_ easier because auto-deriving Ord for GADTs
often gives "inaccessible code" error messages and writing them manually is a
pain in the ass.