Add Mirrors datatype
This is the last of the TUF types to be encoded. In this commit we just define
the datatype, we don't do anything else yet. We need to add infrastructure for
a mirrors role, modify the local tool to sign the mirrors file, and then of
course add support for actually using the mirrors information.