Get rid of the combined snapshot/timestamp file
This was no longer feasible when we start adding out-of-index targets (such as
collections). It's not that useful anyway: mirror clients, or anyone else that
wants to check for updates, can just download the timestamp file, which is
tiny. The snapshot only needs to be downloaded if indeed there are changes.