Avoid committing to a specific enc/dec monad
Some of the TUF datatypes had ToJSON/FromJSON instances defined only for the
specific ReadJSON monad, but this made differenting between different kinds of
effects difficult. By introducing a MonadKeys monad we can avoid this. We can
take advantage of this in clients to not to have to pass in a layout when
writing certain files.

(This also fixes the example client)