Use atomic file operations
Previous commit was not quite right. We can _only_ be sure that we can use
`renameFile` if the two files live in the same directory; in any other
situation the other directory may live on a different physical device in which
case `renameFile` may not work.

In this commit we fix this, and consistently use atomic file operations
throughout.

Fixes #50 (again) and Closes #21.