Do full download if update fails with IOException
We were doing this before, but then we changed it to retry an incremental
download on an exception the first time round the verification loop. However,
that's not right -- if this is the first time round the verification loop, but
the error we get on an incremental download is an IOException (or any other
error apart from a verification error) then we should do a full download
instead.