Start working on integration tests (Issue #67)
For now we test a successful run only: check for updates twice starting with an
empty cache, verify that the first returns HasUpdates and the next returns
NoUpdates. We need to start introducing errors, but the infrastructure is in
place for that sort of thing now: we have a mock-up in-memory cache and repo so
that we can start introducing intentional errors.