Treat deserialization errs as verification errs
This is important, because some verification errors will materialize as
deserialization errors. For example, if we rollover the timestamp key, then
deserialization of the timestamp will fail. If we don't regard this as a
verification error, we will not go through the download-new-root-and-try-again
loop and the check for updates will fail. Closes #108.

This also Closes #75, because although we didn't add a source to
DeserializationError, we now through this through

   VerificationErrorDeserialization TargetPath DeserializationError

which _does_ have the `TargetPath`.