On a full overwrite of 00-index.tar, fully rebuild index
If another program overwrites the 00-index.tar then we rewrite it in
full rather than just appending. This patch also makes sure that we
fully rebuild the .tar.idx index file too in this situation. This is
not strictly necesary in the typical case but it's sensible to do to
avoid corner cases.