Change JSON conversion to ignore unknown Map keys rather than failing
In general we ignore object fields that we don't expect, but the
instance for Map would always fail for keys it could not understand.
This is a problem for forwards compatability. In particular for hash
functions we want to be able to accept (and ignore) new hash function
names.

FromObjectKey is the class of types that can be used as keys in Maps.
This patch changes fromObjectKey to return Maybe. Most of the instances
cannot fail anyway, so the changes are trivial. The HashFn instance is
changed to return Nothing for unknown hash function names. Similarly,
the instance for TargetPath returns Nothing for unknown prefixes.