In file content verification, ignore md5 hashes
The file content verification works by taking the known trusted file
info that we expect (from the TUF metadata files) and the actual file
info from computing the hashes of the target file and simply comparing
them. The file info contains a Map HashFn Hash, ie an entry for multiple
hash function algorithms. So the problem is that if the TUF metadata
contained both md5 and sha256 hashes but we only computed the sha256 of
the target file then these two maps to not compare equal.

So now the code has a clearer explicit policy on which hash functions we
will check and which ones we will allow but ignore. Currently that
policy is that we check sha256 and ignore md5.