Add more sanity checks to incremental update of uncompressed index.tar
The current code blindly assumes that the pre-existing index.tar is a
prefix to the new uncompressed index.tar;

Besides the possibility of data corruption, the incremental
index.tar.gz update logic however supports update transactions which
can violate this assumption, resulting in a state where the index.tar
doesn't get updated or alternative gets corrupted for real.

This patch makes the heuristic for the incremental update of the
uncompressed index.tar more robust by including more sanity checks,
specifically testing that the tail of the old index.tar is contained
in the new index.tar data stream it is to be updated with. If these
new integrity checks fail, the code falls back to the (slower)
non-incremental full decompression codepath.

This is only a partial solution to #196.