Report missing .meta keys as SchemaErrors
This change was triggered by trying make the codebase MonadFail
compatible which exposed a `lookupM` convenience wrapper which
reported lookup failures via Monad(fail). However, schema errors
such as missing keys or similar are supposed to be reported via
the `ReportSchemaError(expected)` mechanism.

This however prompts a major version increment to 0.6.0.0.

So this fixes two issues with one commit.