Use blocking file locking for cache lock
Previously the cache lock operation
(Hackage.Security.Client.Repository.Cache.lockCache) would try once to
lock the cache directory and fail if the lock was already held.
This leads to spurious failures in `cabal-install` due to its concurrent
package fetching behavior.

In this case it's perfectly sensible to rather block.