Add support for logging cache locking events
Since we can now block for indefinite amount of time if we can't get
the cache lock it's useful to have some sort of logging output to
diagnose a stuck lock more easily if this ever occurs; moreover in
combination with `-vverbose+timestamp` it allows us to see when and
for how long the repo cache lock is held during cabal operations which
is also quite useful information to have.

This extends the `LogMessage` type with additional constructors to
allow to encode events for cache lock acquiring and release, and adds
a call-back to the (non-public) `withDirLock` for emitting the
respective events; A new (public) `lockCacheWithLogger` function is
added which makes use of this new `withDirLock` callback.